requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
python-telegram-bot==13.13
pytz==2024.1
requests==2.32.3
six==1.16.0
tornado==6.4.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
